requirements.txt
pillow
opencv-python
watchdog
